move component-submission submission wizard page to correct location
