+++ packages/component-submission/client/pages/SubmissionWizard/ProgressBar.js
