+++ packages/component-submission/client/pages/SubmissionWizard/WizardStep.js
