+++ packages/component-submission/client/pages/SubmissionWizard/formDataModifiers/EditorSuggestionsModifier.js
