+++ packages/component-submission/client/pages/SubmissionWizard/formDataModifiers/EditorSuggestionsModifier.test.js
