+++ packages/component-submission/client/pages/SubmissionWizard/steps/Author/AuthorPage.js
