+++ packages/component-submission/client/pages/SubmissionWizard/steps/Author/index.js
