+++ packages/component-submission/client/pages/SubmissionWizard/steps/Author/schema.js
