+++ packages/component-submission/client/pages/SubmissionWizard/steps/Disclosure/schema.js
