+++ packages/component-submission/client/pages/SubmissionWizard/steps/Editors/EditorsPage.md
