+++ packages/component-submission/client/pages/SubmissionWizard/steps/Editors/index.js
