+++ packages/component-submission/client/pages/SubmissionWizard/steps/Editors/schema.js
