+++ packages/component-submission/client/pages/SubmissionWizard/steps/Files/CoverLetterEditor.js
