+++ packages/component-submission/client/pages/SubmissionWizard/steps/Files/ManuscriptUpload.test.js
