+++ packages/component-submission/client/pages/SubmissionWizard/steps/Files/index.test.js
