+++ packages/component-submission/client/pages/SubmissionWizard/steps/Files/schema.js
