+++ packages/component-submission/client/pages/SubmissionWizard/steps/Files/utils.js
