+++ packages/component-submission/client/pages/SubmissionWizard/steps/Submission/schema.js
